+++ Questionable/Windows/Common/Ui/QstWidgets.cs
@@ -4,7 +4,6 @@ using Dalamud.Bindings.ImGui;
 using Dalamud.Interface;
 using Dalamud.Interface.Utility;
 using Dalamud.Interface.Utility.Raii;
-using Questionable.Utils;
 
 namespace Questionable.Windows.Common.Ui;
 
@@ -133,19 +132,23 @@ internal static class QstWidgets
     }
 
     // Icon button with a tooltip and an optional count badge.
-    public static bool RailButton(FontAwesomeIcon icon, string tooltip, Vector4? tint = null,
-        bool enabled = true, string? countBadge = null, Vector4? badgeColor = null,
+    public static bool RailButton(FontAwesomeIcon icon, string label, string? tooltip = null, Vector4? tint = null,
+        bool enabled = true, string? countBadge = null, Vector4? badgeColor = null, bool showLabel = false,
         [CallerFilePath] string file = "", [CallerLineNumber] int line = 0)
     {
         float size = ImGui.GetFrameHeight();
         bool clicked;
         using (ImRaii.Disabled(!enabled))
         {
-            clicked = ImGuiComponentsLocal.IconButton(icon, tint, activeColor: null, hoveredColor: null,
-                new Vector2(size, size), file, line);
+            if (showLabel)
+                clicked = ImGuiComponentsLocal.IconButtonWithText(icon, label, tint, activeColor: null, hoveredColor: null,
+                    file: file, line: line);
+            else
+                clicked = ImGuiComponentsLocal.IconButton(icon, tint, activeColor: null, hoveredColor: null,
+                    new Vector2(size, size), file, line);
         }
 
-        if (countBadge != null)
+        if (!showLabel && countBadge != null)
         {
             Vector2 max = ImGui.GetItemRectMax();
             Vector2 badgeSize = ImGui.CalcTextSize(countBadge);
@@ -155,8 +158,14 @@ internal static class QstWidgets
 
         if (ImGui.IsItemHovered(ImGuiHoveredFlags.AllowWhenDisabled))
         {
-            using ImRaii.IEndObject _ = ImRaii.Tooltip();
-            ImGui.TextUnformatted(tooltip);
+            if (showLabel && tooltip?.Length != 0)
+                ImGui.SetTooltip(tooltip);
+            else
+            {
+                if (tooltip?.Length != 0)
+                    tooltip = $"\n{tooltip}";
+                ImGui.SetTooltip($"{label}{tooltip}");
+            }
         }
 
         return clicked && enabled;