+++ Questionable/Windows/QuestComponents/CreationUtilsComponent.cs
@@ -1,4 +1,4 @@
-using Dalamud.Bindings.ImGui;
+﻿using Dalamud.Bindings.ImGui;
 using Dalamud.Game.ClientState.Conditions;
 using Dalamud.Game.ClientState.Objects.Types;
 using Dalamud.Game.Text;
@@ -13,8 +13,8 @@ using FFXIVClientStructs.FFXIV.Client.Game.UI;
 using Questionable.Model.Common;
 using Questionable.Model.Questing;
 using Dalamud.Game.ClientState.Objects;
-using ObjectKind = Dalamud.Game.ClientState.Objects.Enums.ObjectKind;
 using Questionable.Windows.Common.Ui;
+using ObjectKind = Dalamud.Game.ClientState.Objects.Enums.ObjectKind;
 
 namespace Questionable.Windows.QuestComponents;
 
@@ -193,22 +193,25 @@ internal sealed class CreationUtilsComponent
             ImGui.SameLine();
             DrawCopyButton();
         }
-
-        ImGui.SameLine();
-        DrawPathEditorButton();
     }
 
-    private void DrawPathEditorButton()
+    internal void DrawPathEditorButton(ElementId? currentQuest, bool sameLine = false)
     {
-        ElementId? currentQuest = questFunctions.GetCurrentQuest().CurrentQuest;
         using (ImRaii.Disabled(currentQuest == null))
         {
-            if (ImGuiComponentsLocal.IconButton(FontAwesomeIcon.Edit) && currentQuest != null)
-                pathEditorWindow.Open(currentQuest);
+            if (currentQuest != null)
+            {
+                if (sameLine)
+                    ImGui.SameLine();
+                if (ImGuiComponentsLocal.IconButton(FontAwesomeIcon.Edit))
+                    pathEditorWindow.Open(currentQuest);
+                else if (ImGui.IsItemClicked(ImGuiMouseButton.Right))
+                    questRegistry.OpenEditor(currentQuest.Value);
+            }
         }
 
         if (ImGui.IsItemHovered(ImGuiHoveredFlags.AllowWhenDisabled))
-            ImGui.SetTooltip(_L("Open the current quest in the Path Editor."));
+            ImGui.SetTooltip(_L("Left click: Open in Path Editor\nRight click: Open in your default .json text editor"));
     }
 
     private unsafe void DrawTargetDetails(IGameObject target)
@@ -339,9 +342,8 @@ internal sealed class CreationUtilsComponent
         return $"{q.CurrentQuest} → {q.Sequence} - {qw}";
     }
 
-    private unsafe void DrawCopyButton(IGameObject target)
+    private void DrawCopyButton(IGameObject target)
     {
-        GameObject* gameObject = (GameObject*)target.Address;
         bool copy = ImGuiComponentsLocal.IconButton(FontAwesomeIcon.Copy);
         if (ImGui.IsItemHovered())
         {